Move git classes to app module
