+++ app/src/main/scala/giter8/ConsoleCredentialsProvider.scala
