+++ app/src/main/scala/giter8/JgitHelper.scala
