+++ app/src/test/scala/giter8/JGitInteractorTest.scala
