+++ library/src/main/scala/giter8/Giter8.scala
