+++ library/src/main/scala/giter8/TestFileHelpers.scala
@@ -1,11 +1,25 @@
 package giter8
 
+/*
+ * Original implementation (C) 2010-2015 Nathan Hamblen and contributors
+ * Adapted and extended in 2016 by foundweekends project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import java.io.{File, PrintWriter}
 
 import org.apache.commons.io.FileUtils
-import org.scalatest.matchers.{MatchResult, Matcher}
-
-import scala.io.Source
 
 trait TestFileHelpers {
   def tempDirectory(code: File => Unit): Unit = {
@@ -33,18 +47,6 @@ trait TestFileHelpers {
         write(s)
         close()
       }
-    }
-  }
-
-  def haveContents(contents: String): Matcher[File] = new Matcher[File] {
-    override def apply(left: File): MatchResult = {
-      val fileContents = Source.fromFile(left).getLines().mkString("\n")
-      MatchResult(
-        matches = fileContents == contents,
-        rawFailureMessage =
-          s"File ${left.getAbsolutePath} has invalid contents:\n\texpected $contents\n\tactual: $fileContents",
-        rawNegatedFailureMessage = s"${left.getAbsolutePath} has contents $contents"
-      )
     }
   }
 }
