+++ library/src/test/scala/giter8/FileRendererTest.scala
@@ -3,7 +3,7 @@ package giter8
 import org.codehaus.plexus.components.io.attributes.PlexusIoResourceAttributeUtils
 import org.scalatest.{FlatSpec, Matchers}
 
-class FileRendererTest extends FlatSpec with Matchers with TestFileHelpers {
+class FileRendererTest extends FlatSpec with Matchers with TestFileHelpers with FileContentMatchers {
   import FileDsl._
   import FileRenderer._
